pyproject.toml
[project]
name = "tarea3-pdc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
]
